Do rendering entirely on the GPU
The javascript now only sends the slice number and the plane, sending
the intensity data itself to the GPU as a 3d texture.

This avoids an x*y loop on the javascript side to set the intensity
data for every pixel displayed, significantly improving performance.

However, it limits the files that can be loaded to ones that are below
the maximum texture size of the OpenGL driver.

Also resolves #14